feat: Refactored closed chats rooms table filters
