+++ src/views/chats/ClosedChats/RoomsTable.vue
@@ -1,44 +1,6 @@
 <template>
   <section class="closed-chats__rooms-table">
-    <rooms-table-filters-loading v-if="isFiltersLoading" />
-    <section v-else class="closed-chats__rooms-table__handlers">
-      <div class="closed-chats__rooms-table__handlers__input">
-        <unnnic-label :label="$t('chats.search_contact')" />
-        <unnnic-input
-          v-model="filterContact"
-          icon-left="search-1"
-          :placeholder="$t('name_or_phone')"
-        />
-      </div>
-      <div class="closed-chats__rooms-table__handlers__input" v-if="sectorsToFilter.length > 2">
-        <unnnic-label :label="$t('sector.title')" />
-        <unnnic-select-smart v-model="filterSector" :options="sectorsToFilter" ordered-by-index />
-      </div>
-      <div class="closed-chats__rooms-table__handlers__input">
-        <unnnic-label :label="$t('tags.title')" />
-        <unnnic-select-smart
-          v-model="filterTag"
-          :disabled="filterSector[0]?.value === 'all' || tagsToFilter.length < 2"
-          :options="tagsToFilter"
-          multiple
-        />
-      </div>
-      <div class="closed-chats__rooms-table__handlers__input">
-        <unnnic-label :label="$t('date')" />
-        <unnnic-input-date-picker
-          class="closed-chats__rooms-table__handlers__date-picker"
-          v-model="filterDate"
-          position="right"
-          :inputFormat="$t('date_format')"
-        />
-      </div>
-      <unnnic-button
-        :text="$t('clear')"
-        :disabled="isFiltersDefault"
-        type="secondary"
-        @click="resetFilters"
-      />
-    </section>
+    <closed-chats-rooms-table-filters @update-filters="filters = $event" />
 
     <rooms-table-loading v-if="isTableLoading" />
     <unnnic-table
@@ -101,22 +63,20 @@
 </template>
 
 <script>
-import moment from 'moment';
-
-import Sector from '@/services/api/resources/settings/sector';
 import History from '@/services/api/resources/chats/history';
 
-import RoomsTableFiltersLoading from '@/views/loadings/ClosedChats/RoomsTableFiltersLoading';
 import RoomsTableLoading from '@/views/loadings/ClosedChats/RoomsTableLoading';
 import TablePagination from '@/components/TablePagination';
 import TagGroup from '@/components/TagGroup.vue';
+
+import ClosedChatsRoomsTableFilters from './RoomsTableFilters.vue';
 
 export default {
   name: 'ClosedChatsRoomsTable',
 
   components: {
     TagGroup,
-    RoomsTableFiltersLoading,
+    ClosedChatsRoomsTableFilters,
     RoomsTableLoading,
     TablePagination,
   },
@@ -129,7 +89,6 @@ export default {
   },
 
   data: () => ({
-    isFiltersLoading: true,
     isTableLoading: true,
     isPagesLoading: true,
 
@@ -139,44 +98,15 @@ export default {
     roomsCurrentPage: 1,
     roomsLimit: 5,
 
-    sectorsToFilter: [],
-    tagsToFilter: [],
-    filterContact: '',
-    filterSector: [],
-    filterTag: [],
-    filterDate: {
-      start: moment().subtract(1, 'week').format('YYYY-MM-DD'),
-      end: moment().format('YYYY-MM-DD'),
+    filters: {
+      contact: '',
+      sector: [],
+      tag: [],
+      date: null,
     },
   }),
 
-  async created() {
-    this.isFiltersLoading = true;
-
-    this.filterSector = [this.filterSectorsOptionAll];
-    this.filterDate = this.filterDateDefault;
-    this.tagsToFilter = this.filterTagDefault;
-    await this.getSectors();
-
-    this.isFiltersLoading = false;
-  },
-
   computed: {
-    filterSectorsOptionAll() {
-      return { value: 'all', label: this.$t('all') };
-    },
-
-    filterTagDefault() {
-      return [{ value: '', label: this.$t('filter.by_tags') }];
-    },
-
-    filterDateDefault() {
-      return {
-        start: moment().subtract(1, 'week').format('YYYY-MM-DD'),
-        end: moment().format('YYYY-MM-DD'),
-      };
-    },
-
     tableHeaders() {
       return [
         {
@@ -206,28 +136,6 @@ export default {
         },
       ];
     },
-
-    isFiltersDefault() {
-      const {
-        sectorsToFilter,
-        filterContact,
-        filterSector,
-        filterTag,
-        filterDate,
-        filterDateDefault,
-      } = this;
-
-      if (
-        filterContact === '' &&
-        (filterSector[0]?.value === 'all' || sectorsToFilter.length === 2) &&
-        filterTag.length === 0 &&
-        filterDate === filterDateDefault
-      ) {
-        return true;
-      }
-
-      return false;
-    },
   },
 
   methods: {
@@ -235,24 +143,27 @@ export default {
       this.isTableLoading = true;
       this.isPagesLoading = true;
 
-      const { roomsCurrentPage, roomsLimit, filterDate, filterContact, filterSector, filterTag } =
-        this;
+      const {
+        roomsCurrentPage,
+        roomsLimit,
+        filters: { date, contact, sector, tag },
+      } = this;
 
       if (paginate !== true) {
         this.roomsCurrentPage = 1;
       }
 
       const offset = (roomsCurrentPage - 1) * roomsLimit;
-      const tagsToReq = filterTag.map((tag) => tag.label).join(',');
-      const sectionToReq = filterSector[0]?.value === 'all' ? '' : filterSector[0]?.value;
+      const tagsToReq = tag.map((tag) => tag.label).join(',');
+      const sectionToReq = sector[0]?.value === 'all' ? '' : sector[0]?.value;
 
       try {
         const response = await History.getHistoryRooms({
           offset,
           limit: roomsLimit,
-          ended_at_before: filterDate.end,
-          ended_at_after: filterDate.start,
-          search: filterContact,
+          ended_at_before: date.end,
+          ended_at_after: date.start,
+          search: contact,
           sector: sectionToReq,
           tag: tagsToReq,
         });
@@ -266,72 +177,13 @@ export default {
       this.isTableLoading = false;
       this.isPagesLoading = false;
     },
-
-    async getSectors() {
-      try {
-        const { results } = await Sector.list();
-
-        const newSectors = [this.filterSectorsOptionAll];
-        results.forEach(({ uuid, name }) => newSectors.push({ value: uuid, label: name }));
-        this.sectorsToFilter = newSectors;
-
-        if (results.length === 1) {
-          this.filterSector = [newSectors[1]];
-        }
-
-        if (results.length > 0) {
-          this.getSectorTags(results[0].uuid);
-        }
-      } catch (error) {
-        console.error('The sectors could not be loaded at this time.', error);
-      }
-    },
-
-    async getSectorTags(sectorUuid) {
-      if (!sectorUuid) {
-        this.tagsToFilter = [];
-        return;
-      }
-      try {
-        const { results } = await Sector.tags(sectorUuid);
-
-        const newTags = this.tagsToFilter;
-        results.forEach(({ uuid, name }) => newTags.push({ value: uuid, label: name }));
-        this.tagsToFilter = newTags;
-      } catch (error) {
-        console.error('The sector tags could not be loaded at this time.', error);
-      }
-    },
-
-    resetFilters() {
-      if (this.isFiltersDefault) {
-        return;
-      }
-
-      this.filterContact = '';
-      if (this.sectorsToFilter.length > 2) {
-        this.filterSector = [this.filterSectorsOptionAll];
-      }
-      this.filterTag = [];
-      this.filterDate = this.filterDateDefault;
-    },
   },
 
   watch: {
     roomsCurrentPage() {
       this.getHistoryRooms(true);
     },
-    filterContact() {
-      const TIME_TO_WAIT_TYPING = 800;
-
-      if (this.timeout !== 0) clearTimeout(this.timeout);
-      this.timeout = setTimeout(() => {
-        this.getHistoryRooms();
-      }, TIME_TO_WAIT_TYPING);
-    },
-    filterSector: 'getHistoryRooms',
-    filterTag: 'getHistoryRooms',
-    filterDate: 'getHistoryRooms',
+    filters: 'getHistoryRooms',
   },
 };
 </script>
@@ -347,28 +199,6 @@ export default {
 
     grid-template-rows: auto 1fr auto;
     height: 100%;
-
-    &__handlers {
-      display: grid;
-      grid-template-columns: repeat(4, 1fr) auto;
-
-      gap: $unnnic-spacing-sm;
-      justify-content: space-between;
-      align-items: flex-end;
-
-      .unnnic-label__label {
-        margin: 0;
-        margin-bottom: $unnnic-spacing-nano;
-      }
-
-      &__input {
-        width: 100%;
-      }
-
-      &__date-picker {
-        display: grid;
-      }
-    }
 
     &__table {
       overflow: hidden;
